[fix] Sanitize strings by removing lone surrogates
