+++ lib/browser.js
@@ -2,7 +2,7 @@
  * Module dependencies.
  */
 
-var keys = require('./keys');
+var utils = require('./utils');
 var hasBinary = require('has-binary');
 var sliceBuffer = require('arraybuffer.slice');
 var after = require('after');
@@ -56,7 +56,7 @@ var packets = exports.packets = {
   , noop:     6
 };
 
-var packetslist = keys(packets);
+var packetslist = utils.keys(packets);
 
 /**
  * Premade error packet.
@@ -117,7 +117,7 @@ exports.encodePacket = function (packet, supportsBinary, utf8encode, callback) {
 
   // data fragment is optional
   if (undefined !== packet.data) {
-    encoded += utf8encode ? utf8.encode(String(packet.data)) : String(packet.data);
+    encoded += utf8encode ? utf8.encode(utils.sanitizeString(String(packet.data))) : String(packet.data);
   }
 
   return callback('' + encoded);
