+++ lib/index.js
@@ -4,7 +4,7 @@
 
 var utf8 = require('utf8');
 var after = require('after');
-var keys = require('./keys');
+var utils = require('./utils');
 
 /**
  * Current protocol version.
@@ -25,7 +25,7 @@ var packets = exports.packets = {
   , noop:     6
 };
 
-var packetslist = keys(packets);
+var packetslist = utils.keys(packets);
 
 /**
  * Premade error packet.
@@ -72,7 +72,7 @@ exports.encodePacket = function (packet, supportsBinary, utf8encode, callback) {
 
   // data fragment is optional
   if (undefined !== packet.data) {
-    encoded += utf8encode ? utf8.encode(String(packet.data)) : String(packet.data);
+    encoded += utf8encode ? utf8.encode(utils.sanitizeString(String(packet.data))) : String(packet.data);
   }
 
   return callback('' + encoded);
